- Fix SolrSpark - Lots of renaming and cleanup
