+++ src/main/scala/cs848/ca/uwaterloo/cs848/HdfsSpark.scala
@@ -1,26 +1,11 @@
-package cs848.nlp
+package ca.uwaterloo.cs848
 
+import ca.uwaterloo.cs848.conf.HdfsConf
+import ca.uwaterloo.cs848.util.{SentenceDetector, Stemmer}
 import org.apache.log4j.Logger
-import org.apache.hadoop.fs.{FileSystem, Path}
 import org.apache.spark.{SparkConf, SparkContext}
-import org.rogach.scallop.ScallopConf
-import org.jsoup.Jsoup
-import cs848.util.{SentenceDetector, Stemmer}
 
-class HDFSConf(args: Seq[String]) extends ScallopConf(args) {
-  mainOptions = Seq(term, field, debug)
-
-  val term = opt[String](descr = "search term", required=true)
-  val field = opt[String](descr = "search field", required=true)
-  val path = opt[String](descr = "hdfs path", required=true)
-  val debug = opt[Boolean](descr = "debug / print")
-
-  codependent(term, field)
-
-  verify()
-}
-
-object HDFSSpark {
+object HdfsSpark {
 
   val log = Logger.getLogger(getClass.getName)
 
@@ -30,7 +15,7 @@ object HDFSSpark {
     val sc = new SparkContext(conf)
     sc.hadoopConfiguration.set("mapreduce.input.fileinputformat.input.dir.recursive", "true")
 
-    val args = new HDFSConf(argv)
+    val args = new HdfsConf(argv)
 
     val searchTerm = Stemmer.stem(args.term())
     log.info("Search Term: " + searchTerm)
@@ -46,16 +31,16 @@ object HDFSSpark {
 
     // read from HDFS
     log.info("Reading from HDFS")
-//    val hdfs = FileSystem.get(new URI("hdfs://192.168.152.203"), sc.hadoopConfiguration)
-//    val paths = hdfs.listFiles(new Path("/ClueWeb09b/ClueWeb09_English_1"), true)
-//
-//    var docs = Map[String, String]()
-//    while (paths.hasNext) {
-//      val path = paths.next().getPath().toString
-//      val textFile = sc.textFile(path).toString
-//
-//      docs += ("id" -> textFile)
-//    }
+    //    val hdfs = FileSystem.get(new URI("hdfs://192.168.152.203"), sc.hadoopConfiguration)
+    //    val paths = hdfs.listFiles(new Path("/ClueWeb09b/ClueWeb09_English_1"), true)
+    //
+    //    var docs = Map[String, String]()
+    //    while (paths.hasNext) {
+    //      val path = paths.next().getPath().toString
+    //      val textFile = sc.textFile(path).toString
+    //
+    //      docs += ("id" -> textFile)
+    //    }
 
     val sentenceDetector = new SentenceDetector()
