+++ src/main/scala/cs848/ca/uwaterloo/cs848/SolrSpark.scala
@@ -1,0 +1,47 @@
+package ca.uwaterloo.cs848
+
+import ca.uwaterloo.cs848.conf.SolrConf
+import ca.uwaterloo.cs848.util.SentenceDetector
+import com.lucidworks.spark.rdd.SelectSolrRDD
+import org.apache.log4j.{Logger, PropertyConfigurator}
+import org.apache.spark.{SparkConf, SparkContext}
+
+object SolrSpark {
+
+  val log = Logger.getLogger(getClass.getName)
+  PropertyConfigurator.configure("log4j.properties")
+
+  def main(argv: Array[String]) = {
+
+    // Parse command line args
+    val args = new SolrConf(argv)
+    log.info(args.summary)
+
+    // Setup Spark
+    val conf = new SparkConf().setAppName(getClass.getSimpleName)
+    val sc = new SparkContext(conf)
+
+    // Start timing the experiment
+    val start = System.currentTimeMillis()
+
+    val (solr, index, rows, field, term, debug) = (args.solr(), args.index(), args.rows(), args.field(), args.term(), args.debug())
+
+    val rdd = new SelectSolrRDD(solr, index, sc)
+      .rows(rows)
+      .query(field + ":" + term)
+      .foreach(doc => {
+        val sentenceDetector = new SentenceDetector()
+        val sentences = sentenceDetector.inference(doc.get(field).toString, field)
+        if (debug) {
+          log.info("ID: " + doc.get("id"))
+          sentences.foreach(println)
+        }
+      })
+
+    // Need to manually call stop()
+    sc.stop()
+
+    log.info(s"Took ${System.currentTimeMillis() - start}ms")
+
+  }
+}
