+++ src/main/scala/cs848/ca/uwaterloo/cs848/util/Stemmer.scala
@@ -1,4 +1,4 @@
-package cs848.util
+package ca.uwaterloo.cs848.util
 
 import org.apache.lucene.analysis.CharArraySet
 import org.apache.lucene.analysis.en.EnglishAnalyzer
